requirements.txt
omegaconf
cdsapi
pandas
matplotlib
numpy
pillow
xarray
netcdf4
scipy
dvc
dask[diagnostics]